Refactored the user select
