+++ src/components/nav/NavBar.js
@@ -2,8 +2,8 @@ import React, { useContext, useEffect, useState } from "react";
 import { ReactComponent as Logo } from "../../man-holding-resume-1-cropped.svg";
 import { useHistory } from "react-router-dom";
 import { UserContext } from "../users/UserProvider";
-import { UserOptions } from "../users/UserOptions";
 import "./NavBar.css";
+import { UserSelect } from "../users/UserSelect";
 
 export const NavBar = (props) => {
   const { currentUser, getCurrentUser } = useContext(UserContext);
@@ -15,18 +15,14 @@ export const NavBar = (props) => {
   const history = useHistory();
   const gotoHomepage = () => history.push("/");
   const gotoCurrentUser = () => history.push(`/resume/${currentUser.id}`);
-  const logout = () => {
-    localStorage.removeItem("trb_user");
-    history.push("/");
-  };
-
-  const [searchTerm, setSearchTerm] = useState("");
-  const changeSearchTerm = (event) => setSearchTerm(event.target.value);
-
-  const handleSelectUser = (event) => {
+  const gotoSelectedUser = (event) => {
     if (event.target.value) {
       history.push(`/resume/${event.target.value}`);
     }
+  };
+  const logout = () => {
+    localStorage.removeItem("trb_user");
+    history.push("/");
   };
 
   return (
@@ -35,19 +31,7 @@ export const NavBar = (props) => {
         <Logo />
       </div>
       <div className="navbar__item">
-        <form>
-          <input
-            type="text"
-            name="filterText"
-            placeholder="Filter by name..."
-            value={searchTerm}
-            onChange={changeSearchTerm}
-          />
-        </form>
-        <select className="navbar__item userSelect" onClick={handleSelectUser}>
-          <option value="">View another resume...</option>
-          <UserOptions searchTerm={searchTerm} />
-        </select>
+        <UserSelect selectFunc={gotoSelectedUser} />
       </div>
       <div className="navbar__item currentUser" onClick={gotoCurrentUser}>
         {currentUser.name}
